Extract status to component
