+++ app/next-client-app/app/scanreports/columns.tsx
@@ -27,6 +27,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { ChevronRight } from "lucide-react";
+import { ScanReportStatus } from "@/components/scanreports/ScanReportStatus";
 
 export const columns: ColumnDef<ScanReportResult>[] = [
   {
@@ -101,43 +102,7 @@ export const columns: ColumnDef<ScanReportResult>[] = [
     enableHiding: false,
     enableSorting: false,
     cell: ({ row }) => {
-      const statusMapping = {
-        BLOCKED: { text: "Blocked", color: "text-red-900" },
-        COMPLET: { text: "Mapping Complete", color: "text-green-600" },
-        INPRO25: { text: "Mapping 25%", color: "text-orange-300" },
-        INPRO50: { text: "Mapping 50%", color: "text-orange-400" },
-        INPRO75: { text: "Mapping 75%", color: "text-orange-500" },
-        UPCOMPL: { text: "Upload Complete", color: "text-blue-800" },
-        UPFAILE: { text: "Upload Failed", color: "text-red-500" },
-        UPINPRO: { text: "Upload in Progress", color: "text-orange-600" },
-      };
-
-      const { id, status } = row.original;
-      const statusInfo = statusMapping[status as keyof typeof statusMapping];
-      const textColorClassName = statusInfo?.color ?? "text-black";
-
-      const handleChangeStatus = async (value: string) => {
-        try {
-          await updateScanReport(id, "status", value);
-        } catch (error) {
-          console.error(error);
-        }
-      };
-
-      return (
-        <Select value={status} onValueChange={handleChangeStatus}>
-          <SelectTrigger className={`${textColorClassName} w-[180px]`}>
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {Object.entries(statusMapping).map(([value, { text, color }]) => (
-              <SelectItem key={value} value={value}>
-                {text}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      );
+      return <ScanReportStatus row={row} />;
     },
   },
   {
